third_party/blink/web_tests/platform/mac-mac14-arm64/virtual/webgl-extra-video-texture-metadata/fast/webgl/video-metadata/texImage-video-last-uploaded-metadata-expected.txt

FAIL Unexpected error: Uncaught TypeError: Cannot read properties of null (reading 'createTexture')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

 
Test FAILED
